Split SSL handshake tests
